% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_squid.R, R/lowest_squid.R
\name{highest_squid}
\alias{highest_squid}
\alias{lowest_squid}
\title{Finding extreme (highest or lowest) SQUIDs}
\usage{
highest_squid(x)

lowest_squid(x)
}
\arguments{
\item{x}{A vector of SQUIDs (or a list of vectors, which will be
recursively \code{\link[=unlist]{unlist()}}ed).}
}
\value{
The highest or lowest SQUID
}
\description{
Finding extreme (highest or lowest) SQUIDs
}
\examples{
squids::highest_squid(
  squids::squids(5)
);

squids::lowest_squid(
  squids::squids(5)
);
}
